Move the notes api methods into a controller in the api namespace
