+++ app/views/api/notes/_entry.html.erb
