+++ app/views/api/notes/_note.gpx.builder
