+++ app/views/api/notes/index.json.jsonify
