+++ app/views/api/notes/show.gpx.builder
