+++ app/views/api/notes/show.xml.builder
